[Feat] ViewModel 반영
