--- NAVER-MAP-iOS/Screen/SearchResult/SearchResultViewModel.swift
@@ -1,8 +1,0 @@
-import UIKit
-
-final class SearchResultViewModel {
-    
-    //MARK: - Properties
-    
-    
-}
